requirements.txt
spotipy
python-dotenv
SpeechRecognition
PyAudio
pillow
sounddevice
numpy
requests
openai-whisper